move Wallets components around
